pyproject.toml
[project]
name = "website-classification"
version = "0.1.0"
description = "Website visual style classification using CLIP + HDBSCAN + Claude VLM"
requires-python = ">=3.10"
dependencies = [
    "playwright>=1.40",
    "open-clip-torch>=2.24",
    "torch>=2.1",
    "Pillow>=10.0",
    "numpy>=1.24",
    "umap-learn>=0.5",
    "hdbscan>=0.8",
    "scikit-learn>=1.3",
    "anthropic>=0.39",
    "python-dotenv>=1.0",
    "tqdm>=4.66",
    "pandas>=2.1",
    "rich>=13.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
